Revamped the data fixtures
